multimodal-rag/requirements.txt
# Core RAG Dependencies
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
instructor>=1.0.0

# Vector Databases (install only what you need)
chromadb>=0.4.22
lancedb>=0.5.0
# qdrant-client>=1.7.0
# weaviate-client>=4.4.0
# pymilvus>=2.3.0

# Embeddings
openai>=1.10.0
sentence-transformers>=2.3.0
transformers>=4.36.0
torch>=2.1.0
pillow>=10.2.0
open-clip-torch>=2.24.0

# Document Processing
unstructured[all-docs]>=0.12.0
pdf2image>=1.16.3
pytesseract>=0.3.10
pypdf>=4.0.0
python-docx>=1.1.0
python-pptx>=0.6.23
markdown>=3.5.0
beautifulsoup4>=4.12.0
lxml>=5.1.0

# Image Processing
opencv-python>=4.9.0
scikit-image>=0.22.0

# Reranking
cohere>=4.47

# UI Frameworks
streamlit>=1.31.0
gradio>=4.16.0

# Evaluation
ragas>=0.1.0
datasets>=2.16.0

# Utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
loguru>=0.7.2
tqdm>=4.66.0
pandas>=2.2.0
numpy>=1.26.0
redis>=5.0.0
httpx>=0.26.0

# Monitoring
prometheus-client>=0.19.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
